python: ecl.faults moved to ecl.grid.faults.
...
