+++ python/ecl/grid/ecl_region.py
@@ -33,7 +33,7 @@ from ecl.util import monkey_the_camel
 from ecl.util import IntVector
 
 from ecl import EclPrototype
-from ecl.faults import Layer
+from ecl.grid.faults import Layer
 from ecl import EclDataType
 from ecl.eclfile import EclKW
 from ecl.geo import CPolyline
