+++ python/ecl/grid/faults/fault.py
