+++ python/ecl/grid/faults/fault_block.py
