+++ python/ecl/grid/faults/fault_block_collection.py
