+++ python/ecl/grid/faults/fault_block_layer.py
@@ -20,7 +20,7 @@ from cwrap import BaseCClass
 from ecl.util import monkey_the_camel
 from ecl import EclDataType
 from ecl import EclPrototype
-from ecl.faults import Fault
+from ecl.grid.faults import Fault
 
 class FaultBlockLayer(BaseCClass):
     TYPE_NAME = "fault_block_layer"
