+++ python/ecl/grid/faults/fault_collection.py
