+++ python/ecl/grid/faults/fault_line.py
