+++ python/ecl/grid/faults/fault_segments.py
