+++ python/ecl/grid/faults/layer.py
