+++ python/tests/ecl_tests/test_faults.py
@@ -22,7 +22,7 @@ from ecl import util
 from ecl import EclDataType
 from ecl.eclfile import EclKW
 from ecl.grid import EclGrid
-from ecl.faults import FaultCollection, Fault, FaultLine, FaultSegment,FaultBlockLayer
+from ecl.grid.faults import FaultCollection, Fault, FaultLine, FaultSegment,FaultBlockLayer
 from ecl.test import TestAreaContext
 from ecl.geo import Polyline , CPolyline
 from tests import EclTest
