+++ python/tests/ecl_tests/test_grid.py
@@ -25,7 +25,7 @@ from ecl import EclDataType, EclUnitTypeEnum
 from ecl.eclfile import EclKW, EclFile
 from ecl.grid import EclGrid
 from ecl.grid import EclGridGenerator as GridGen
-from ecl.faults import Layer , FaultCollection
+from ecl.grid.faults import Layer , FaultCollection
 from ecl.test import TestAreaContext
 from tests import EclTest
